backend/requirements.txt
google-generativeai>=0.8.0
google-cloud-bigquery>=3.19.0
google-cloud-firestore>=2.16.0
google-cloud-pubsub>=2.21.0
google-cloud-scheduler>=2.13.0
google-cloud-monitoring>=2.22.0
google-cloud-secret-manager>=2.20.0
google-auth>=2.29.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.1
pytest>=8.2.0
pytest-asyncio>=0.23.0
ruff>=0.4.0
